feat(BaseLink)!: add missing info about breaking change
BREAKING CHANGE: The `chip-clicked` event has been removed for type chip links.
Chips are now rendered as an `a` tag and a query param with the chip-object is added to the link.
The URL params can be set with the new props `chipQueryName`, `identifierPropertyName` and `path`.

Commit [c1b20dfe] introduced a `breaking change`, but was not marked as having it.

Refs: #2431